requirements-prod.txt
# Production environment dependencies
# Core dependencies only for production deployment

# Django core
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
asgiref==3.9.1
sqlparse==0.5.3

# Database
mysqlclient==2.2.7

# Data processing
pandas==2.2.2
numpy==2.0.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0

# Azure cloud services
azure-storage-blob==12.20.0
azure-core==1.30.1
azure-identity==1.15.0
azure-common==1.1.28

# Encryption and security
cryptography==43.0.3
pycryptodome==3.10.1
PyJWT==2.8.0

# HTTP and networking
requests==2.32.3
urllib3==2.1.0

# Timezone handling
pytz==2023.3.post1
tzdata==2024.1

# Utility libraries
click==8.1.7
packaging==23.2
typing_extensions==4.12.2

# Production environment additional dependencies
gunicorn==21.2.0
whitenoise==6.5.0 

django-cors-headers==4.7.0